backend/fastapi/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
PyJWT==2.9.*
requests==2.32.*
python-dotenv==1.1.*
google-generativeai==0.8.*
psycopg2-binary==2.9.*
sqlalchemy==2.0.*

